fixed a segfault in heredoc

+++ src/utils/memory_utils.c
@@ -6,7 +6,7 @@
 /*   By: yabarhda <yabarhda@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/31 02:43:22 by yabarhda          #+#    #+#             */
-/*   Updated: 2025/08/08 18:21:50 by yabarhda         ###   ########.fr       */
+/*   Updated: 2025/08/09 20:44:09 by yabarhda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,10 @@ void	clean_up(t_token *token)
 	while (token)
 	{
 		if (token->type == REDIR_HEREDOC)
-			unlink(token->next->value);
+		{
+			if (token->next)
+				unlink(token->next->value);
+		}
 		token = token->next;
 	}
 }
